Spinoff driver into own project with demorun.
