+++ samples/attachment-demo/build.gradle
@@ -30,7 +30,7 @@ dependencies {
     cordaCompile project(path: ":webserver:webcapsule", configuration: 'runtimeArtifacts')
     cordaCompile project(':core')
     cordaCompile project(':webserver')
-    cordaCompile project(':test-utils')
+    cordaCompile project(':driver')
 }
 
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
