+++ samples/network-visualiser/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'us.kirchmeier.capsule'
 
 dependencies {
     compile project(':samples:irs-demo')
-    compile project(':test-utils')
+    compile project(':driver')
 
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
     testCompile "junit:junit:$junit_version"
