+++ samples/notary-demo/build.gradle
@@ -23,8 +23,7 @@ dependencies {
     cordaCompile project(':core')
     cordaCompile project(':client:jfx')
     cordaCompile project(':client:rpc')
-    cordaCompile project(':test-utils')
-    cordaCompile project(':cordform-common')
+    cordaCompile project(':driver')
 }
 
 idea {
