+++ samples/simm-valuation-demo/build.gradle
@@ -56,7 +56,7 @@ dependencies {
     compile "com.opengamma.strata:strata-math:${strata_version}"
 
     // Test dependencies
-    testCompile project(':test-utils')
+    testCompile project(':driver')
     testCompile "junit:junit:$junit_version"
     testCompile "org.assertj:assertj-core:${assertj_version}"
 }
