+++ settings.gradle
@@ -21,6 +21,7 @@ include 'verifier'
 include 'test-common'
 include 'test-utils'
 include 'smoke-test-utils'
+include 'driver'
 include 'tools:explorer'
 include 'tools:explorer:capsule'
 include 'tools:demobench'
