+++ test-utils/src/main/kotlin/net/corda/testing/CoreTestUtils.kt
@@ -3,34 +3,26 @@
 
 package net.corda.testing
 
-import com.nhaarman.mockito_kotlin.spy
-import com.nhaarman.mockito_kotlin.whenever
 import net.corda.core.contracts.StateRef
 import net.corda.core.crypto.*
 import net.corda.core.identity.Party
 import net.corda.core.identity.PartyAndCertificate
-import net.corda.core.node.ServiceHub
 import net.corda.core.node.services.IdentityService
-import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.utilities.*
+import net.corda.core.utilities.NetworkHostAndPort
+import net.corda.core.utilities.OpaqueBytes
+import net.corda.core.utilities.loggerFor
 import net.corda.finance.contracts.asset.DUMMY_CASH_ISSUER
-import net.corda.node.services.config.NodeConfiguration
-import net.corda.node.services.config.VerifierType
 import net.corda.node.services.config.configureDevKeyAndTrustStores
 import net.corda.node.services.identity.InMemoryIdentityService
 import net.corda.node.utilities.CertificateType
 import net.corda.node.utilities.X509Utilities
 import net.corda.nodeapi.config.SSLConfiguration
 import net.corda.nodeapi.internal.serialization.AMQP_ENABLED
-import net.corda.testing.node.MockServices
-import net.corda.testing.node.MockServices.Companion.makeTestDataSourceProperties
-import net.corda.testing.node.MockServices.Companion.makeTestDatabaseProperties
 import org.bouncycastle.asn1.x500.X500Name
 import org.bouncycastle.asn1.x500.X500NameBuilder
 import org.bouncycastle.asn1.x500.style.BCStyle
-import java.net.URL
 import java.nio.file.Files
-import java.nio.file.Path
 import java.security.KeyPair
 import java.security.PublicKey
 import java.security.cert.CertificateFactory
@@ -124,62 +116,6 @@ fun freePort(): Int = freePortCounter.getAndAccumulate(0) { prev, _ -> 30000 + (
 fun getFreeLocalPorts(hostName: String, numberToAlloc: Int): List<NetworkHostAndPort> {
     val freePort =  freePortCounter.getAndAccumulate(0) { prev, _ -> 30000 + (prev - 30000 + numberToAlloc) % 10000 }
     return (freePort .. freePort + numberToAlloc - 1).map { NetworkHostAndPort(hostName, it) }
-}
-
-/**
- * Creates and tests a ledger built by the passed in dsl. The provided services can be customised, otherwise a default
- * of a freshly built [MockServices] is used.
- */
-@JvmOverloads fun ledger(
-        services: ServiceHub = MockServices(),
-        initialiseSerialization: Boolean = true,
-        dsl: LedgerDSL<TestTransactionDSLInterpreter, TestLedgerDSLInterpreter>.() -> Unit
-): LedgerDSL<TestTransactionDSLInterpreter, TestLedgerDSLInterpreter> {
-    if (initialiseSerialization) initialiseTestSerialization()
-    try {
-        val ledgerDsl = LedgerDSL(TestLedgerDSLInterpreter(services))
-        dsl(ledgerDsl)
-        return ledgerDsl
-    } finally {
-        if (initialiseSerialization) resetTestSerialization()
-    }
-}
-
-/**
- * Creates a ledger with a single transaction, built by the passed in dsl.
- *
- * @see LedgerDSLInterpreter._transaction
- */
-@JvmOverloads fun transaction(
-        transactionLabel: String? = null,
-        transactionBuilder: TransactionBuilder = TransactionBuilder(notary = DUMMY_NOTARY),
-        initialiseSerialization: Boolean = true,
-        dsl: TransactionDSL<TransactionDSLInterpreter>.() -> EnforceVerifyOrFail
-) = ledger(initialiseSerialization = initialiseSerialization) {
-    dsl(TransactionDSL(TestTransactionDSLInterpreter(this.interpreter, transactionBuilder)))
-}
-
-fun testNodeConfiguration(
-        baseDirectory: Path,
-        myLegalName: X500Name): NodeConfiguration {
-    abstract class MockableNodeConfiguration : NodeConfiguration // Otherwise Mockito is defeated by val getters.
-    val nc = spy<MockableNodeConfiguration>()
-    whenever(nc.baseDirectory).thenReturn(baseDirectory)
-    whenever(nc.myLegalName).thenReturn(myLegalName)
-    whenever(nc.minimumPlatformVersion).thenReturn(1)
-    whenever(nc.keyStorePassword).thenReturn("cordacadevpass")
-    whenever(nc.trustStorePassword).thenReturn("trustpass")
-    whenever(nc.rpcUsers).thenReturn(emptyList())
-    whenever(nc.dataSourceProperties).thenReturn(makeTestDataSourceProperties(myLegalName.commonName))
-    whenever(nc.database).thenReturn(makeTestDatabaseProperties())
-    whenever(nc.emailAddress).thenReturn("")
-    whenever(nc.exportJMXto).thenReturn("")
-    whenever(nc.devMode).thenReturn(true)
-    whenever(nc.certificateSigningService).thenReturn(URL("http://localhost"))
-    whenever(nc.certificateChainCheckPolicies).thenReturn(emptyList())
-    whenever(nc.verifierType).thenReturn(VerifierType.InMemory)
-    whenever(nc.messageRedeliveryDelaySeconds).thenReturn(5)
-    return nc
 }
 
 @JvmOverloads
