+++ tools/explorer/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     compile project(':core')
     compile project(':client:jfx')
     compile project(':client:mock')
-    compile project(':test-utils')
+    compile project(':driver')
     compile project(':finance')
 
     // Capsule is a library for building independently executable fat JARs.
