+++ verifier/build.gradle
@@ -39,7 +39,7 @@ dependencies {
     compile "org.apache.logging.log4j:log4j-slf4j-impl:${log4j_version}"
     compile "org.apache.logging.log4j:log4j-core:${log4j_version}"
 
-    integrationTestCompile project(":test-utils")
+    integrationTestCompile project(":driver")
     integrationTestCompile project(":client:mock")
 
     // Integration test helpers
