+++ webserver/build.gradle
@@ -57,7 +57,7 @@ dependencies {
     // For rendering the index page.
     compile "org.jetbrains.kotlinx:kotlinx-html-jvm:0.6.3"
 
-    testCompile project(':test-utils')
+    integrationTestCompile project(':driver')
     testCompile "junit:junit:$junit_version"
 }
